+++ workers/bubbleSort/run.py
@@ -2,35 +2,17 @@
 import websockets
 import asyncio
 import json
+from bubbleSort import bubbleSort
 
 connection = websockets.connect(uri='ws://localhost:8765')
 
 MSG_TYPE = {
-    "CONNECT": 1,
+    "REGISTER": 1,
     "REQUEST": 2,
     "RESPONSE": 3
 }
 
 IS_REGISTERED = False
-
-def bubble_sort(arr):
-    # Taken from here: https://www.geeksforgeeks.org/python-program-for-bubble-sort/
-    
-    # Outer loop to iterate through the list n times
-    for n in range(len(arr) - 1, 0, -1):
-        swapped = False  
-        for i in range(n):
-            if arr[i] > arr[i + 1]:
-
-                # Swap elements if they are in the wrong order
-                arr[i], arr[i + 1] = arr[i + 1], arr[i]
-                swapped = True
-        
-        if not swapped:
-            break
-
-    return arr
-    
 
 async def main():
     async with connection as websocket:
@@ -39,7 +21,7 @@ async def main():
         async for message in websocket:
             message = json.loads(message)
 
-            if message["code"] == MSG_TYPE["CONNECT"]:
+            if message["code"] == MSG_TYPE["REGISTER"]:
                 global IS_REGISTERED
                 IS_REGISTERED = True
             else:
