+++ workers/mergeSort/mergeSort.py
@@ -1,16 +1,3 @@
-
-import websockets
-import asyncio
-import json
-
-connection = websockets.connect(uri='ws://localhost:8765')
-
-MSG_TYPE = {
-    "CONNECT": 1,
-    "REQUEST": 2,
-    "RESPONSE": 3
-}
-
 def merge(arr, l, m, r):
     '''
         Reference: https://www.geeksforgeeks.org/python-program-for-merge-sort/
@@ -77,29 +64,3 @@ def mergeSort(arr, l, r):
         merge(arr, l, m, r)
 
     return arr
-
-
-async def main():
-    async with connection as websocket:
-        await websocket.send('{"code": 1, "worker": true, "type": "mergeSort"}')
-
-        async for message in websocket:
-            message = json.loads(message)
-
-            if message["code"] == MSG_TYPE["CONNECT"]:
-                global IS_REGISTERED
-                IS_REGISTERED = True
-            else:
-                print(f"Task: {message['type']}, Value:{message['data']}")
-                n = len(message["data"])
-                resp = {
-                    "code": MSG_TYPE["RESPONSE"],
-                    "worker": True,
-                    "type": "mergeSort",
-                    "value": mergeSort(message["data"], 0, n-1)
-                }
-                await websocket.send(json.dumps(resp))
-            
-        await websocket.close()
-
-asyncio.run(main())
